pyproject.toml
[tool.poetry]
name = "sqr-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
passlib = "1.7.4"
pyjwt = "2.8.0"
requests = "2.31.0"
streamlit = "1.33.0"
sqlalchemy = "1.4.52"
fastapi = "0.110.2"
uvicorn = "^0.29.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
email-validator = "^2.1.1"
fastapi-login = "^1.10.0"
pytest = "^8.1.2"
bandit = "^1.7.8"
pytest-cov = "^5.0.0"
mutmut = "^2.4.5"
mock = "^5.1.0"
flake8 = "^7.0.0"
httpx = "^0.27.0"
extra-streamlit-components = "^0.1.71"
locust = "^2.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
